RUBY-1033: Implement find, get more, kill cursors, explain commands for 3.2

+++ spec/mongo/command_monitoring_spec.rb
@@ -1,47 +1,60 @@
 require 'spec_helper'
 
-def runnable?(file)
+def skippable?(file)
   !write_command_enabled? && (file.include?('bulkWrite') || file.include?('insert'))
+end
+
+def ignore?(test)
+  if version = test.ignore_if_server_version_greater_than
+    return true if version == "3.0" && find_command_enabled?
+  end
+  if version = test.ignore_if_server_version_less_than
+    return true if version == "3.1" && !find_command_enabled?
+  end
+  false
 end
 
 describe 'Command Monitoring Events' do
 
   COMMAND_MONITORING_TESTS.each do |file|
 
-    spec = Mongo::CommandMonitoring::Spec.new(file)
+    if !skippable?(file)
 
-    spec.tests.each do |test|
+      spec = Mongo::CommandMonitoring::Spec.new(file)
 
-      context(test.description) do
+      spec.tests.each do |test|
 
-        let(:subscriber) do
-          Mongo::CommandMonitoring::TestSubscriber.new
-        end
+        context(test.description) do
 
-        let(:monitoring) do
-          authorized_client.instance_variable_get(:@monitoring)
-        end
+          let(:subscriber) do
+            Mongo::CommandMonitoring::TestSubscriber.new
+          end
 
-        before do
-          authorized_collection.find.delete_many
-          authorized_client.subscribe(Mongo::Monitoring::COMMAND, subscriber)
-        end
+          let(:monitoring) do
+            authorized_client.instance_variable_get(:@monitoring)
+          end
 
-        after do
-          monitoring.subscribers[Mongo::Monitoring::COMMAND].delete(subscriber)
-          authorized_collection.find.delete_many
-        end
+          before do
+            authorized_collection.find.delete_many
+            authorized_client.subscribe(Mongo::Monitoring::COMMAND, subscriber)
+          end
 
-        test.expectations.each do |expectation|
+          after do
+            monitoring.subscribers[Mongo::Monitoring::COMMAND].delete(subscriber)
+            authorized_collection.find.delete_many
+          end
 
-          it "generates a #{expectation.event_name} for #{expectation.command_name}", unless: runnable?(file) do
-            begin
-              test.run(authorized_collection)
-              event = subscriber.send(expectation.event_type)[expectation.command_name]
-              expect(event).to send(expectation.matcher, expectation)
-            rescue Mongo::Error::OperationFailure, Mongo::Error::BulkWriteError => e
-              event = subscriber.send(expectation.event_type)[expectation.command_name]
-              expect(event).to send(expectation.matcher, expectation)
+          test.expectations.each do |expectation|
+
+            it "generates a #{expectation.event_name} for #{expectation.command_name}", unless: ignore?(test) do
+              begin
+                test.run(authorized_collection)
+                event = subscriber.send(expectation.event_type)[expectation.command_name]
+                expect(event).to send(expectation.matcher, expectation)
+              rescue Mongo::Error::OperationFailure, Mongo::Error::BulkWriteError => e
+                event = subscriber.send(expectation.event_type)[expectation.command_name]
+                expect(event).to send(expectation.matcher, expectation)
+              end
             end
           end
         end
